requirements.txt
aiohttp==3.8.3
asyncpg==0.27.0
discord.py==2.3.1
fastapi==0.83.0
python-dotenv==0.19.1
requests==2.26.0